backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Security and Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Validation and Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Environment Variables
python-dotenv>=1.0.0

# AI & Machine Learning
openai>=1.3.0
langchain>=0.0.340
langchain-openai>=0.0.2

# Data Processing
pandas>=2.1.0
numpy>=1.25.0

# Azure Services
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
requests>=2.31.0

# Production Server
gunicorn>=21.2.0

# Performance & Caching
redis>=5.0.0
aioredis>=2.0.0

# File Processing
Pillow>=10.1.0
openpyxl>=3.1.0

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2023.3
aiosmtplib>=3.0.0